+++ Sources/BraintreeSEPADirectDebit/BTSEPADirectDebitClient.swift
@@ -10,7 +10,8 @@ import BraintreeCore
 
     // MARK: - Internal Properties
 
-    let apiClient: BTAPIClient
+    // Exposed for unit tests
+    var apiClient: BTAPIClient
     
     var webAuthenticationSession: BTWebAuthenticationSession
         
@@ -19,10 +20,10 @@ import BraintreeCore
     // MARK: - Initializers
 
     ///  Creates a SEPA Direct Debit client.
-    /// - Parameter apiClient: An instance of `BTAPIClient`
-    @objc(initWithAPIClient:)
-    public init(apiClient: BTAPIClient) {
-        self.apiClient = apiClient
+    /// - Parameter authorization: A valid client token or tokenization key used to authorize API calls
+    @objc(initWithAuthorization:)
+    public init(authorization: String) {
+        self.apiClient = BTAPIClient(newAuthorization: authorization)
         self.sepaDirectDebitAPI = SEPADirectDebitAPI(apiClient: apiClient)
         self.webAuthenticationSession = BTWebAuthenticationSession()
 
@@ -31,8 +32,8 @@ import BraintreeCore
     }
     
     /// Internal for testing.
-    init(apiClient: BTAPIClient, webAuthenticationSession: BTWebAuthenticationSession, sepaDirectDebitAPI: SEPADirectDebitAPI) {
-        self.apiClient = apiClient
+    init(authorization: String, webAuthenticationSession: BTWebAuthenticationSession, sepaDirectDebitAPI: SEPADirectDebitAPI) {
+        self.apiClient = BTAPIClient(newAuthorization: authorization)
         self.webAuthenticationSession = webAuthenticationSession
         self.sepaDirectDebitAPI = sepaDirectDebitAPI
     }