+++ UnitTests/BraintreeSEPADirectDebitTests/BTSEPADirectDebitClient_Tests.swift
@@ -9,9 +9,10 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
     var billingAddress = BTPostalAddress()
     var sepaDirectDebitRequest = BTSEPADirectDebitRequest()
     var mockAPIClient : MockAPIClient = MockAPIClient(authorization: "development_client_key")!
+    let authorization: String = "sandbox_9dbg82cq_dcpspy2brwdjr3qn"
 
     override func setUp() {
-        mockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
+        mockAPIClient = MockAPIClient(authorization: authorization)!
 
         billingAddress.streetAddress = "Kantstraße 70"
         billingAddress.extendedAddress = "#170"
@@ -29,7 +30,9 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
     }
 
     func testTokenizeWithPresentationContext_callsCreateMandateWithError_returnsError_andSendsAnalytics() {
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(apiClient: mockAPIClient)
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization)
+        sepaDirectDebitClient.apiClient = mockAPIClient
+        
         let mockSepaDirectDebitAPI = SEPADirectDebitAPI(apiClient: mockAPIClient)
         mockAPIClient.cannedResponseError = NSError(
             domain: "CannedError",
@@ -72,10 +75,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
         )
 
         let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+            authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         sepaDirectDebitClient.tokenize(sepaDirectDebitRequest) { nonce, error in
             if error != nil, let error = error as NSError? {
@@ -111,11 +116,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
 
         mockWebAuthenticationSession.cannedResponseURL = URL(string: "invalid-url")
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         sepaDirectDebitClient.tokenize(sepaDirectDebitRequest) { nonce, error in
             if error != nil, let error = error as NSError? {
@@ -139,11 +145,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
             userInfo: ["Description": "Mock noBodyReturned error description."]
         )
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
         
         mockAPIClient.cannedResponseBody = nil
 
@@ -179,11 +186,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
             ]
         )
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         sepaDirectDebitClient.tokenize(sepaDirectDebitRequest) { nonce, error in
             if error != nil, let error = error as NSError? {
@@ -219,11 +227,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
 
         mockWebAuthenticationSession.cannedResponseURL = nil
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         sepaDirectDebitClient.tokenize(sepaDirectDebitRequest) { nonce, error in
             if error != nil, let error = error as NSError? {
@@ -257,11 +266,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
             ]
         )
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         let mockTokenizeResponse = BTJSON(
             value: [
@@ -321,11 +331,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
             ] as [String: Any]
         )
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         mockAPIClient.cannedResponseBody = mockCreateMandateResult
         mockAPIClient.cannedResponseBody = mockTokenizeResponse
@@ -370,11 +381,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
             userInfo: [NSLocalizedDescriptionKey: "This is a fake tokenize request error"]
         )
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         sepaDirectDebitClient.tokenize(sepaDirectDebitRequest) { nonce, error in
             if error != nil, let error = error as NSError? {
@@ -415,11 +427,12 @@ class BTSEPADirectDebitClient_Tests: XCTestCase {
             userInfo: [NSLocalizedDescriptionKey: "This is a fake tokenize request error"]
         )
 
-        let sepaDirectDebitClient = BTSEPADirectDebitClient(
-            apiClient: mockAPIClient,
+        let sepaDirectDebitClient = BTSEPADirectDebitClient(authorization: authorization,
             webAuthenticationSession: mockWebAuthenticationSession,
             sepaDirectDebitAPI: mockSepaDirectDebitAPI
         )
+        
+        sepaDirectDebitClient.apiClient = mockAPIClient
 
         sepaDirectDebitClient.tokenize(sepaDirectDebitRequest) { nonce, error in
             if error != nil, let error = error as NSError? {