+++ UnitTests/BraintreeSEPADirectDebitTests/SEPADirectDebitAPI_Tests.swift
@@ -8,6 +8,7 @@ class SEPADirectDebitAPI_Tests: XCTestCase {
     var sepaDirectDebitRequest = BTSEPADirectDebitRequest()
     var successApprovalURL: String = ""
     var mockAPIClient : MockAPIClient = MockAPIClient(authorization: "development_client_key")!
+    let authorization: String = "sandbox_9dbg82cq_dcpspy2brwdjr3qn"
     var mockCreateMandateResult = CreateMandateResult(json:
         BTJSON(
             value: [
@@ -54,7 +55,6 @@ class SEPADirectDebitAPI_Tests: XCTestCase {
     }
 
     func testCreateMandate_properlyFormatsPOSTBody() {
-        let sepaDirectDebitRequest = BTSEPADirectDebitRequest()
         billingAddress.streetAddress = "fake-street-addres"
         billingAddress.extendedAddress = "fake-extended-address"
         billingAddress.locality = "fake-locality"
@@ -67,7 +67,16 @@ class SEPADirectDebitAPI_Tests: XCTestCase {
         sepaDirectDebitRequest.billingAddress = billingAddress
         sepaDirectDebitRequest.merchantAccountID = "fake-account-id"
         sepaDirectDebitRequest.locale = "fr-FR"
-        
+
+        let sepaDirectDebitRequest = BTSEPADirectDebitRequest(
+            accountHolderName: "fake-name",
+            iban: "fake-iban",
+            customerID: "fake-customer-id",
+            billingAddress: billingAddress,
+            merchantAccountID: "fake-account-id",
+            locale: "fr-FR"
+        )
+
         let api = SEPADirectDebitAPI(apiClient: mockAPIClient)
         api.createMandate(sepaDirectDebitRequest: sepaDirectDebitRequest) { _, _ in }